Make all filter exceptions consistent
Even though there is no test case that can make the previous code in
SearchApi fail, it is more future-proof to get the parameter name from
the exception thrower rather than $key. If InvalidAndOperatorException,
InvalidNotOperatorException or InvalidFilterUsageException are ever to be
thrown from a filter's compile(), the error message will include the
correct parameter name.
