+++ src/Model/Exception/InvalidNotOperatorException.php
@@ -4,6 +4,6 @@ namespace App\Model\Exception;
 
 use Exception;
 
-class InvalidNotOperatorException extends Exception
+class InvalidNotOperatorException extends FilterException
 {
 }
