+++ src/Model/Exception/InvalidValueException.php
@@ -4,17 +4,6 @@ namespace App\Model\Exception;
 
 use Exception;
 
-class InvalidValueException extends Exception
+class InvalidValueException extends FilterException
 {
-    protected $thrower;
-
-    public function __construct() {
-        $args = func_get_args();
-        $this->thrower = array_shift($args);
-        parent::__construct(...$args);
-    }
-
-    public function getThrower() {
-        return $this->thrower;
-    }
 }
